Conflicts:
	TFC API/TFC/API/Events/PlayerSkillEvent.java
	TFC API/TFC/API/IBag.java
	TFC Resources/assets/terrafirmacraft/textures/items/Calamari Cooked.png
	TFC Resources/assets/terrafirmacraft/textures/items/Calamari Raw.png
	TFC Resources/assets/terrafirmacraft/textures/items/beefCooked.png
	TFC Resources/assets/terrafirmacraft/textures/items/beefRaw.png
	TFC Resources/assets/terrafirmacraft/textures/items/chickenCooked.png
	TFC Resources/assets/terrafirmacraft/textures/items/chickenRaw.png
	TFC Resources/assets/terrafirmacraft/textures/items/fishCooked.png
	TFC Resources/assets/terrafirmacraft/textures/items/fishRaw.png
	TFC Resources/assets/terrafirmacraft/textures/items/food/Calamari Cooked.png
	TFC Resources/assets/terrafirmacraft/textures/items/food/Calamari Raw.png
	TFC Resources/assets/terrafirmacraft/textures/items/food/beefCooked.png
	TFC Resources/assets/terrafirmacraft/textures/items/food/beefRaw.png
	TFC Resources/assets/terrafirmacraft/textures/items/food/chickenCooked.png
	TFC Resources/assets/terrafirmacraft/textures/items/food/chickenRaw.png
	TFC Resources/assets/terrafirmacraft/textures/items/food/fishCooked.png
	TFC Resources/assets/terrafirmacraft/textures/items/food/fishRaw.png
	TFC Resources/assets/terrafirmacraft/textures/items/food/porkchopCooked.png
	TFC Resources/assets/terrafirmacraft/textures/items/food/porkchopRaw.png
	TFC Resources/assets/terrafirmacraft/textures/items/porkchopCooked.png
	TFC Resources/assets/terrafirmacraft/textures/items/porkchopRaw.png
	TFC_Shared/src/TFC/Core/Player/FoodStatsTFC.java
	TFC_Shared/src/TFC/Core/Player/PlayerTracker.java
	TFC_Shared/src/TFC/Core/Player/SkillStats.java
	TFC_Shared/src/TFC/Core/TFC_Textures.java
	TFC_Shared/src/TFC/Food/CropIndex.java
	TFC_Shared/src/TFC/Food/CropIndexPepper.java
	TFC_Shared/src/TFC/Food/CropManager.java
	TFC_Shared/src/TFC/Food/ItemMeal.java
	TFC_Shared/src/TFC/Handlers/CraftingHandler.java
	TFC_Shared/src/TFC/Handlers/PacketHandler.java
	TFC_Shared/src/TFC/Items/ItemBlocks/ItemFoodBlock.java
	TFC_Shared/src/TFC/TFCItems.java
	TFC_Shared/src/TFC/TileEntities/TileEntityFoodPrep.java
	TFC_Shared/src/TFC/TileEntities/TileEntityIngotPile.java
	TFC_Shared/src/TFC/TileEntities/TileEntityLogPile.java
	TFC_Shared/src/TFC/TileEntities/TileEntityPottery.java
	TFC_Shared/src/TFC/TileEntities/TileEntityToolRack.java
	TFC_Shared/src/TFC/TileEntities/TileEntityWorkbench.java
	api/TFC/API/Constant/Global.java
	src/main/java/TFC/Blocks/BlockCrop.java
	src/main/java/TFC/Blocks/BlockPlanks.java
	src/main/java/TFC/Blocks/BlockPlanks2.java
	src/main/java/TFC/Blocks/BlockTerra.java
	src/main/java/TFC/Containers/ContainerFoodPrep.java
	src/main/java/TFC/Containers/ContainerVessel.java
	src/main/java/TFC/Containers/Slots/SlotFoodOnly.java
	src/main/java/TFC/Core/Recipes.java
	src/main/java/TFC/Core/TFC_Core.java
	src/main/java/TFC/Core/TFC_ItemHeat.java
	src/main/java/TFC/Entities/Mobs/EntityChickenTFC.java
	src/main/java/TFC/Entities/Mobs/EntityCowTFC.java
	src/main/java/TFC/Entities/Mobs/EntityDeer.java
	src/main/java/TFC/Entities/Mobs/EntityPheasantTFC.java
	src/main/java/TFC/Entities/Mobs/EntityPigTFC.java
	src/main/java/TFC/Entities/Mobs/EntitySheepTFC.java
	src/main/java/TFC/Entities/Mobs/EntitySquidTFC.java
	src/main/java/TFC/GUI/GuiCalendar.java
	src/main/java/TFC/GUI/GuiContainerTFC.java
	src/main/java/TFC/GUI/GuiFoodPrep.java
	src/main/java/TFC/GUI/GuiInventoryTFC.java
	src/main/java/TFC/GUI/GuiSkills.java
	src/main/java/TFC/Handlers/AnvilCraftingHandler.java
	src/main/java/TFC/Handlers/EntityLivingHandler.java
	src/main/java/TFC/Items/ItemBlocks/ItemTerraBlock.java
	src/main/java/TFC/Items/ItemTerra.java
	src/main/java/TFC/Items/Pottery/ItemPotterySmallVessel.java
	src/main/java/TFC/TerraFirmaCraft.java
	src/main/java/TFC/TileEntities/TECrop.java
	src/main/java/TFC/TileEntities/TileEntityAnvil.java
	src/main/java/TFC/TileEntities/TileEntityBarrel.java
	src/main/java/TFC/TileEntities/TileEntityChestTFC.java
	src/main/java/TFC/TileEntities/TileEntityFirepit.java
	src/main/java/TFC/TileEntities/TileEntityForge.java
	src/main/java/TFC/TileEntities/TileEntityQuern.java
	src/main/resources/assets/terrafirmacraft/lang/en_US.properties
	src/main/resources/assets/terrafirmacraft/textures/items/Calamari Cooked.png
	src/main/resources/assets/terrafirmacraft/textures/items/Calamari Raw.png
	src/main/resources/assets/terrafirmacraft/textures/items/beefCooked.png
	src/main/resources/assets/terrafirmacraft/textures/items/beefRaw.png
	src/main/resources/assets/terrafirmacraft/textures/items/chickenCooked.png
	src/main/resources/assets/terrafirmacraft/textures/items/chickenRaw.png
	src/main/resources/assets/terrafirmacraft/textures/items/fishCooked.png
	src/main/resources/assets/terrafirmacraft/textures/items/fishRaw.png
	src/main/resources/assets/terrafirmacraft/textures/items/porkchopCooked.png
	src/main/resources/assets/terrafirmacraft/textures/items/porkchopRaw.png

+++ src/main/java/TFC/Core/Player/FoodStatsTFC.java
@@ -241,7 +241,7 @@ public class FoodStatsTFC
 		return this.satisfaction;
 	}
 
-	public void setFoodLevel(int par1)
+	public void setFoodLevel(float par1)
 	{
 		this.stomachLevel = par1;
 	}
